% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valhalla.R
\name{decode}
\alias{decode}
\title{Decode Valhalla Route Shape}
\usage{
decode(encoded)
}
\arguments{
\item{encoded}{An encoded shapefile in ASCII format from Valhalla's API.}
}
\value{
A tibble containing point locations in \code{lat} and \code{lon} columns.
}
\description{
For point-to-point routing, Valhalla's API provides a route shapefile in a
special ASCII-encoded format. This function takes an encoded string, decodes
it, and returns the lat/lon coordinates as a tibble.
}
\details{
To map the results, see also \code{valhallr::map_trip()}.
}
