% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valhalla.R
\name{map_isochrone}
\alias{map_isochrone}
\title{Generate maps of isochrones}
\usage{
map_isochrone(isochrone, method = "leaflet")
}
\arguments{
\item{isochrone}{An isochrone sf object generated by `valhallr::isochrone()`.}

\item{method}{The method used to map it. Two methods are supported:
* "leaflet" produces an interactive HTML map using the Leaflet package.
* "ggplot" produces a static}
}
\value{
A plot of the isochrones, either a a leaflet object or a ggplot object.
}
\description{
Generate maps of isochrones
}
