% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valhalla.R
\name{print_trip}
\alias{print_trip}
\title{Print Trip Summary and Turn-By-Turn Directions}
\usage{
print_trip(trip, all_details = FALSE)
}
\arguments{
\item{all_details}{}
}
\value{

}
\description{
Print Trip Summary and Turn-By-Turn Directions
}
