% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation.R
\name{route_elevation}
\alias{route_elevation}
\title{Title}
\usage{
route_elevation(route, hostname = "localhost", port = 8002)
}
\arguments{
\item{route}{A route object from route.}

\item{hostname}{Hostname or IP address of your Valhalla instance. Defaults to "localhost".}

\item{port}{The port your Valhalla instance is monitoring. Defaults to 8002.}
}
\value{
a sf data.frame with cumulative distance, elevation and point geometry.
}
\description{
Title
}
\examples{
\dontrun{
library(valhallr)

# set up origin and destination data
from = test_data("leeds_university")
to   = test_data("leeds_chemic_pub")

# calculate the trip
trip = route(from, to, costing = "pedestrian")

# get the elevation & cumulative distance
ele = route_elevation(trip)
}

}
