% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valhalla.R
\name{sf_to_latlon}
\alias{sf_to_latlon}
\title{Convert \code{sf} objects to \strong{valhallr}-friendly tibbles}
\usage{
sf_to_latlon(data, output_crs = "WGS84")
}
\arguments{
\item{data}{A simple feature collection with geometry class POINT.}

\item{output_crs}{The desired output coordinate reference system (CRS).
Defaults to WGS84.}
}
\value{
A non-sf tibble with new columns \code{lat} and \code{lon} containing latitudes
and longitudes respectively that can be fed into other \strong{valhallr}
functions like \code{valhallr::od_table()}.
}
\description{
This function converts simple feature (sf) POINT objects to tidy
tibbles that you can feed to \strong{valhallr} functions. This is handy if you're
working with a lots of geospatial data and would like to run a set of
locations through, for example, \code{valhallr::od_table()}.
}
