% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valhalla.R
\name{test_data}
\alias{test_data}
\title{Get Lat/Lon Coordinates for Testing}
\usage{
test_data(dataset = NA)
}
\arguments{
\item{dataset}{The name of a test dataset. By default, and if an unknown input
is given, it returns all values.}
}
\value{
A tibble with one or more location names, latitudes, and longitudes.
}
\description{
This function gives quick access to lat/lon coordinates for a few points
around Ontario for testing and benchmarking purposes.
}
