% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valhalla.R
\name{trace_attributes}
\alias{trace_attributes}
\title{Map Matching with Valhalla}
\usage{
trace_attributes(
  shape = tibble::tibble(lon = numeric(0), lat = numeric(0)),
  costing = "auto",
  unit = "kilometers",
  shape_match = "walk_or_snap",
  begin_time = NULL,
  durations = NULL,
  use_timestamps = FALSE,
  trace_options = list(),
  hostname = "localhost",
  port = 8002
)
}
\arguments{
\item{shape}{A tibble containing point locations in \code{lat} and \code{lon} columns, and optionally a column with timestamps (for GPS data) for speed estimation.}

\item{costing}{The travel costing method. Values "auto", "bicycle", "bus", and "pedestrian" all work.}

\item{unit}{Distance measurement units. Defaults to "kilometres".}

\item{shape_match}{\code{shape_match} is an optional string input parameter. It allows some control of the matching algorithm based on the type of input.}

\item{begin_time}{Begin timestamp for the trace. This is used along with the \code{durations} so that timestamps can be specified for a trace that is specified using an encoded polyline. Options are "edge_walk", "map_snap", and "walk_or_snap". "walk_or_snap" is the default option. See API documentation for more details.}

\item{durations}{List of durations (seconds) between each successive pair of input trace points. This allows trace points to be supplied as an encoded polyline and timestamps to be created by using this list of "delta" times along with the \code{begin_time} of the trace.}

\item{use_timestamps}{A boolean value indicating whether the input timestamps or durations should be used when computing elapsed time at each edge along the matched path. If true, timestamps are used. If false (default), internal costing is applied to compute elapsed times.}

\item{trace_options}{A named list of options provided to the Valhalla API that affect map matching
e.g., search radius. See API documentation for details.}

\item{hostname}{Hostname or IP address of your Valhalla instance. Defaults to "localhost".}

\item{port}{The port your Valhalla instance is monitoring. Defaults to 8002.}
}
\value{
A list with resulting objects from the trace_attributes API. See \href{https://valhalla.readthedocs.io/en/latest/api/map-matching/api-reference/#outputs-of-trace_attributes}{documentation} for more details.
}
\description{
This function calls Valhalla's \code{trace_attributes} API to return a matched series of
OSM ways estimated to represent the route of your trace. Several costing methods are supported, and there are
parameters that let you give custom options to Valhalla. \strong{Please note that this
function requires access to a running instance of Valhalla.}
}
\details{
For more details, please check the Valhalla API documentation here:
\itemize{
\item \url{https://valhalla.readthedocs.io/en/latest/api/map-matching/api-reference/}
}
}
