% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valhalla.R
\name{valhalla_route}
\alias{valhalla_route}
\title{Point-to-Point Routing with Valhalla}
\usage{
valhalla_route(
  from = NA,
  to = NA,
  costing = "auto",
  unit = "kilometers",
  min_road_class = "residential",
  minimum_reachability = 50
)
}
\arguments{
\item{unit}{}
}
\value{

}
\description{
Note: This is a highly experimental function and will almost certainly change.
Right now from and to are one-row tibbles. That's probably not correct.
}
\details{
It does no input validation, no error handling, nothing. Use at own risk.
}
