% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valhalla.R
\name{isochrone}
\alias{isochrone}
\title{Generate isochrones}
\usage{
isochrone(
  from,
  costing = "pedestrian",
  contours_time = c(5, 10, 15),
  min_road_class = "residential",
  minimum_reachability = 500
)
}
\arguments{
\item{minimum_reachability}{}
}
\value{

}
\description{
Generate isochrones
}
