% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valhalla.R
\name{isochrone}
\alias{isochrone}
\title{Generate isochrones}
\usage{
isochrone(
  from,
  costing = "pedestrian",
  contours = c(5, 10, 15),
  metric = "min",
  min_road_class = "residential",
  minimum_reachability = 500
)
}
\arguments{
\item{metric}{Distance or time. Accepts parameters "min" and "km".}

\item{min_road_class}{The minimum road classification Valhalla will consider. Defaults to `residential`.}

\item{minimum_reachability}{The minimum number of nodes a candidate network
needs to have before it is included.}
}
\value{

}
\description{
https://valhalla.readthedocs.io/en/latest/api/isochrone/api-reference/
}
