% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valhalla.R
\name{map_isochrone}
\alias{map_isochrone}
\title{Generate maps of isochrones}
\usage{
map_isochrone(isochrone, method = "leaflet")
}
\arguments{
\item{method}{}
}
\value{

}
\description{
Generate maps of isochrones
}
