% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valhalla.R
\name{map_trip}
\alias{map_trip}
\title{Make a Leaflet Map from a Trip}
\usage{
map_trip(trip, method = "leaflet")
}
\arguments{
\item{method}{Which mapping service to use. Defaults to leaflet; also can use ggplot.}
}
\value{

}
\description{
Make a Leaflet Map from a Trip
}
