% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valhalla.R
\name{print_trip}
\alias{print_trip}
\title{Print Trip Summary and Turn-By-Turn Directions}
\usage{
print_trip(trip, all_details = FALSE)
}
\arguments{
\item{trip}{A trip response from `valhallr::route()`.}

\item{all_details}{Boolean. Should we print each turn-by-turn instruction
along with an overall summary?}
}
\value{
The input `trip` object, invisibly.
}
\description{
Print Trip Summary and Turn-By-Turn Directions
}
