% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valhalla.R
\name{test_data}
\alias{test_data}
\title{Get Lat/Lon Coordinates for Testing}
\usage{
test_data(dataset)
}
\value{
A one-row tibble with a location's name, latitude, and longitude.
}
\description{
This function gives quick access to lat/lon coordinates for a few points
around Ontario for testing and benchmarking purposes.
}
